+++ src/main/java/net/p3pp3rf1y/sophisticatedcore/compat/CompatModIds.java
@@ -12,4 +12,5 @@ public class CompatModIds {
 	public static final String ITEM_BORDERS = "itemborders";*/
 	public static final String CHIPPED = "chipped";
 	public static final String LITEMATICA = "litematica";
+	public static final String AUDIOPLAYER = "audioplayer";
 }