+++ src/main/java/net/p3pp3rf1y/sophisticatedcore/upgrades/jukebox/JukeboxUpgradeItem.java
@@ -6,7 +6,6 @@ import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.RecordItem;
 import net.minecraft.world.level.Level;
@@ -77,13 +76,12 @@ public class JukeboxUpgradeItem extends UpgradeItemBase<JukeboxUpgradeItem.Wrapp
 
 		public void play(Level world, BlockPos pos) {
 			play(world, (serverWorld, storageUuid) ->
-					ServerStorageSoundHandler.startPlayingDisc(serverWorld, pos, storageUuid, Item.getId(getDisc().getItem()), () -> setIsPlaying(false)));
+					ServerStorageSoundHandler.startPlayingDisc(serverWorld, pos, storageUuid, getDisc(), () -> setIsPlaying(false)));
 		}
 
 		public void play(LivingEntity entity) {
 			play(entity.level(), (world, storageUuid) ->
-					ServerStorageSoundHandler.startPlayingDisc(world, entity.position(), storageUuid, entity.getId(),
-							Item.getId(getDisc().getItem()), () -> setIsPlaying(false)));
+					ServerStorageSoundHandler.startPlayingDisc(world, entity.position(), storageUuid, entity.getId(), getDisc(), () -> setIsPlaying(false)));
 		}
 
 		private void play(Level world, BiConsumer<ServerLevel, UUID> play) {