+++ src/main/java/net/p3pp3rf1y/sophisticatedcore/upgrades/jukebox/ServerStorageSoundHandler.java
@@ -1,16 +1,22 @@
 package net.p3pp3rf1y.sophisticatedcore.upgrades.jukebox;
 
+import com.google.common.collect.Lists;
+
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
 import net.p3pp3rf1y.sophisticatedcore.network.PacketHandler;
 
 import java.lang.ref.WeakReference;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.function.Function;
 
 public class ServerStorageSoundHandler {
 	private ServerStorageSoundHandler() {}
@@ -18,6 +24,39 @@ public class ServerStorageSoundHandler {
 	private static final int KEEP_ALIVE_CHECK_INTERVAL = 10;
 	private static final Map<ResourceKey<Level>, Long> lastWorldCheck = new HashMap<>();
 	private static final Map<ResourceKey<Level>, Map<UUID, KeepAliveInfo>> worldStorageSoundKeepAlive = new HashMap<>();
+
+	private static final List<SoundHandler> soundHandlers = Lists.newArrayList();
+
+	public static void registerSoundHandler(SoundHandler handler) {
+		soundHandlers.add(handler);
+	}
+
+	static {
+		registerSoundHandler(new SoundHandler() {
+			@Override
+			public boolean play(ServerLevel level, BlockPos position, UUID storageUuid, ItemStack discItemStack) {
+				Vec3 pos = Vec3.atCenterOf(position);
+				PacketHandler.sendToAllNear(level, pos, 128, new PlayDiscMessage(storageUuid, Item.getId(discItemStack.getItem()), position));
+				return true;
+			}
+
+			@Override
+			public boolean play(ServerLevel level, Vec3 position, UUID storageUuid, int entityId, ItemStack discItemStack) {
+				PacketHandler.sendToAllNear(level, position, 128, new PlayDiscMessage(storageUuid, Item.getId(discItemStack.getItem()), entityId));
+				return true;
+			}
+
+			@Override
+			public void stop(ServerLevel level, Vec3 position, UUID storageUuid) {
+				sendStopMessage(level, position, storageUuid);
+			}
+
+			@Override
+			public void update(UUID storageUuid, Vec3 position) {
+				// noop
+			}
+		});
+	}
 
 	public static void tick(ServerLevel world) {
 		if (world.isClientSide()) {
@@ -31,7 +70,7 @@ public class ServerStorageSoundHandler {
 
 		worldStorageSoundKeepAlive.get(dim).entrySet().removeIf(entry -> {
 			if (entry.getValue().getLastKeepAliveTime() < world.getGameTime() - KEEP_ALIVE_CHECK_INTERVAL) {
-				sendStopMessage(world, entry.getValue().getLastPosition(), entry.getKey());
+				entry.getValue().getSoundHandler().stop(world, entry.getValue().getLastPosition(), entry.getKey());
 				return true;
 			}
 			return false;
@@ -45,7 +84,7 @@ public class ServerStorageSoundHandler {
 			return;
 		}
 		if (worldStorageSoundKeepAlive.get(dim).containsKey(storageUuid)) {
-			worldStorageSoundKeepAlive.get(dim).get(storageUuid).update(world.getGameTime(), position);
+			worldStorageSoundKeepAlive.get(dim).get(storageUuid).update(world.getGameTime(), position, storageUuid);
 		}
 	}
 
@@ -57,11 +96,13 @@ public class ServerStorageSoundHandler {
 		private final WeakReference<Runnable> onStopHandler;
 		private long lastKeepAliveTime;
 		private Vec3 lastPosition;
+		private final SoundHandler handler;
 
-		private KeepAliveInfo(Runnable onStopHandler, long lastKeepAliveTime, Vec3 lastPosition) {
+		private KeepAliveInfo(Runnable onStopHandler, long lastKeepAliveTime, Vec3 lastPosition, SoundHandler handler) {
 			this.onStopHandler = new WeakReference<>(onStopHandler);
 			this.lastKeepAliveTime = lastKeepAliveTime;
 			this.lastPosition = lastPosition;
+			this.handler = handler;
 		}
 
 		public long getLastKeepAliveTime() {
@@ -72,9 +113,14 @@ public class ServerStorageSoundHandler {
 			return lastPosition;
 		}
 
-		public void update(long gameTime, Vec3 position) {
+		public SoundHandler getSoundHandler() {
+			return handler;
+		}
+
+		public void update(long gameTime, Vec3 position, UUID storageUuid) {
 			lastKeepAliveTime = gameTime;
 			lastPosition = position;
+			handler.update(storageUuid, position);
 		}
 
 		public void runOnStop() {
@@ -85,24 +131,34 @@ public class ServerStorageSoundHandler {
 		}
 	}
 
-	public static void startPlayingDisc(ServerLevel serverWorld, BlockPos position, UUID storageUuid, int discItemId, Runnable onStopHandler) {
+	private static void runSoundHandler(ServerLevel serverWorld, Vec3 position, UUID storageUuid, Runnable onStopHandler, Function<SoundHandler, Boolean> onHandler) {
+		for (SoundHandler handler : Lists.reverse(soundHandlers)) {
+			if (onHandler.apply(handler)) {
+				putKeepAliveInfo(serverWorld, storageUuid, onStopHandler, position, handler);
+				return;
+			}
+		}
+	}
+
+	public static void startPlayingDisc(ServerLevel serverWorld, BlockPos position, UUID storageUuid, ItemStack discItemStack, Runnable onStopHandler) {
 		Vec3 pos = Vec3.atCenterOf(position);
-		PacketHandler.sendToAllNear(serverWorld, pos, 128, new PlayDiscMessage(storageUuid, discItemId, position));
-		putKeepAliveInfo(serverWorld, storageUuid, onStopHandler, pos);
+		runSoundHandler(serverWorld, pos, storageUuid, onStopHandler, (handler) -> handler.play(serverWorld, position, storageUuid, discItemStack));
 	}
 
-	public static void startPlayingDisc(ServerLevel serverWorld, Vec3 position, UUID storageUuid, int entityId, int discItemId, Runnable onStopHandler) {
-		PacketHandler.sendToAllNear(serverWorld, position, 128, new PlayDiscMessage(storageUuid, discItemId, entityId));
-		putKeepAliveInfo(serverWorld, storageUuid, onStopHandler, position);
+	public static void startPlayingDisc(ServerLevel serverWorld, Vec3 position, UUID storageUuid, int entityId, ItemStack discItemStack, Runnable onStopHandler) {
+		runSoundHandler(serverWorld, position, storageUuid, onStopHandler, (handler) -> handler.play(serverWorld, position, storageUuid, entityId, discItemStack));
 	}
 
-	private static void putKeepAliveInfo(ServerLevel serverWorld, UUID storageUuid, Runnable onStopHandler, Vec3 pos) {
-		worldStorageSoundKeepAlive.computeIfAbsent(serverWorld.dimension(), dim -> new HashMap<>()).put(storageUuid, new KeepAliveInfo(onStopHandler, serverWorld.getGameTime(), pos));
+	private static void putKeepAliveInfo(ServerLevel serverWorld, UUID storageUuid, Runnable onStopHandler, Vec3 pos, SoundHandler handler) {
+		worldStorageSoundKeepAlive.computeIfAbsent(serverWorld.dimension(), dim -> new HashMap<>()).put(storageUuid, new KeepAliveInfo(onStopHandler, serverWorld.getGameTime(), pos, handler));
 	}
 
 	public static void stopPlayingDisc(ServerLevel serverWorld, Vec3 position, UUID storageUuid) {
+		ResourceKey<Level> dim = serverWorld.dimension();
+		if (worldStorageSoundKeepAlive.containsKey(dim) && worldStorageSoundKeepAlive.get(dim).containsKey(storageUuid)) {
+			worldStorageSoundKeepAlive.get(dim).get(storageUuid).getSoundHandler().stop(serverWorld, position, storageUuid);
+		}
 		removeKeepAliveInfo(serverWorld, storageUuid);
-		sendStopMessage(serverWorld, position, storageUuid);
 	}
 
 	private static void removeKeepAliveInfo(ServerLevel serverWorld, UUID storageUuid) {